+++ src/models/AbstractStyledComponent.js
@@ -2,7 +2,7 @@
 import { Component } from 'react'
 import PropTypes from 'prop-types'
 
-import { CHANNEL } from './ThemeProvider'
+import { CHANNEL, CHANNEL_NEXT, CONTEXT_CHANNEL_SHAPE } from './ThemeProvider'
 import StyleSheet, { CONTEXT_KEY } from './StyleSheet'
 
 export default class AbstractStyledComponent extends Component {
@@ -12,10 +12,19 @@ export default class AbstractStyledComponent extends Component {
     generatedClassName?: string,
     generatedStyles?: any
   }
-  unsubscribe: () => void
+
+  unsubscribeId: number = -1
+
+  unsubscribeFromContext() {
+    const styledContext = this.context[CHANNEL_NEXT]
+    if (styledContext) {
+      styledContext.unsubscribe(this.unsubscribeId)
+    }
+  }
 }
 
 AbstractStyledComponent.contextTypes = {
   [CHANNEL]: PropTypes.func,
+  [CHANNEL_NEXT]: CONTEXT_CHANNEL_SHAPE,
   [CONTEXT_KEY]: PropTypes.instanceOf(StyleSheet),
 }