+++ src/models/StyledComponent.js
@@ -13,7 +13,7 @@ import getComponentName from '../utils/getComponentName'
 import type { RuleSet, Target } from '../types'
 
 import AbstractStyledComponent from './AbstractStyledComponent'
-import { CHANNEL } from './ThemeProvider'
+import { CHANNEL, CHANNEL_NEXT, CONTEXT_CHANNEL_SHAPE } from './ThemeProvider'
 import StyleSheet, { CONTEXT_KEY } from './StyleSheet'
 
 const escapeRegex = /[[\].#*$><+~=|^:(),"'`]/g
@@ -83,9 +83,10 @@ export default (ComponentStyle: Function, constructWithOptions: Function) => {
       // If there is a theme in the context, subscribe to the event emitter. This
       // is necessary due to pure components blocking context updates, this circumvents
       // that by updating when an event is emitted
-      if (this.context[CHANNEL]) {
-        const subscribe = this.context[CHANNEL]
-        this.unsubscribe = subscribe(nextTheme => {
+      const styledContext = this.context[CHANNEL_NEXT]
+      if (styledContext) {
+        const { subscribe } = styledContext
+        this.unsubscribeId = subscribe(nextTheme => {
           // This will be called once immediately
 
           // Props should take precedence over ThemeProvider, which should take precedence over
@@ -120,9 +121,7 @@ export default (ComponentStyle: Function, constructWithOptions: Function) => {
     }
 
     componentWillUnmount() {
-      if (this.unsubscribe) {
-        this.unsubscribe()
-      }
+      this.unsubscribeFromContext()
     }
 
     render() {
@@ -200,6 +199,7 @@ export default (ComponentStyle: Function, constructWithOptions: Function) => {
     class StyledComponent extends ParentComponent {
       static contextTypes = {
         [CHANNEL]: PropTypes.func,
+        [CHANNEL_NEXT]: CONTEXT_CHANNEL_SHAPE,
         [CONTEXT_KEY]: PropTypes.instanceOf(StyleSheet),
       }