+++ src/models/ThemeProvider.js
@@ -6,9 +6,17 @@ import isFunction from 'is-function'
 import isPlainObject from 'is-plain-object'
 import createBroadcast from '../utils/create-broadcast'
 import type { Broadcast } from '../utils/create-broadcast'
+import once from '../utils/once'
 
 // NOTE: DO NOT CHANGE, changing this is a semver major change!
 export const CHANNEL = '__styled-components__'
+export const CHANNEL_NEXT = `${CHANNEL}next__`
+
+export const CONTEXT_CHANNEL_SHAPE = PropTypes.shape({
+  getTheme: PropTypes.func,
+  subscribe: PropTypes.func,
+  unsubscribe: PropTypes.func,
+})
 
 export type Theme = {[key: string]: mixed}
 type ThemeProviderProps = {|
@@ -16,6 +24,11 @@ type ThemeProviderProps = {|
   theme: Theme | (outerTheme: Theme) => void,
 |}
 
+
+const warnChannelDeprecated = once(() => {
+  // eslint-disable-next-line no-console
+  console.error(`Warning: Usage of \`context.${CHANNEL}\` as a function is deprecated. It will be replaced with the object on \`.context.${CHANNEL_NEXT}\` in a future version.`)
+})
 /**
  * Provide a theme to an entire react component tree via context and event listeners (have to do
  * both context and event emitter as pure components block context updates)
@@ -23,9 +36,10 @@ type ThemeProviderProps = {|
 class ThemeProvider extends Component {
   getTheme: (theme?: Theme | (outerTheme: Theme) => void) => Theme
   outerTheme: Theme
-  unsubscribeToOuter: () => void
+  unsubscribeToOuterId: string
   props: ThemeProviderProps
   broadcast: Broadcast
+  unsubscribeToOuterId: number = -1
 
   constructor() {
     super()
@@ -35,9 +49,9 @@ class ThemeProvider extends Component {
   componentWillMount() {
     // If there is a ThemeProvider wrapper anywhere around this theme provider, merge this theme
     // with the outer theme
-    if (this.context[CHANNEL]) {
-      const subscribe = this.context[CHANNEL]
-      this.unsubscribeToOuter = subscribe(theme => {
+    if (this.context[CHANNEL_NEXT]) {
+      const { subscribe } = this.context[CHANNEL_NEXT]
+      this.unsubscribeToOuterId = subscribe(theme => {
         this.outerTheme = theme
       })
     }
@@ -45,7 +59,21 @@ class ThemeProvider extends Component {
   }
 
   getChildContext() {
-    return { ...this.context, [CHANNEL]: this.broadcast.subscribe }
+    return {
+      ...this.context,
+      [CHANNEL_NEXT]: {
+        getTheme: this.getTheme,
+        subscribe: this.broadcast.subscribe,
+        unsubscribe: this.broadcast.unsubscribe,
+      },
+      [CHANNEL]: (subscriber) => {
+        warnChannelDeprecated()
+
+        // Patch the old `subscribe` provide via `CHANNEL` for older clients.
+        const unsubscribeId = this.broadcast.subscribe(subscriber)
+        return () => this.broadcast.unsubscribe(unsubscribeId)
+      },
+    }
   }
 
   componentWillReceiveProps(nextProps: ThemeProviderProps) {
@@ -53,8 +81,9 @@ class ThemeProvider extends Component {
   }
 
   componentWillUnmount() {
-    if (this.context[CHANNEL]) {
-      this.unsubscribeToOuter()
+    const styledContext = this.context[CHANNEL_NEXT]
+    if (styledContext) {
+      styledContext.unsubscribe(this.unsubscribeToOuterId)
     }
   }
 
@@ -83,10 +112,11 @@ class ThemeProvider extends Component {
 }
 
 ThemeProvider.childContextTypes = {
-  [CHANNEL]: PropTypes.func.isRequired,
+  [CHANNEL]: PropTypes.func, // legacy
+  [CHANNEL_NEXT]: CONTEXT_CHANNEL_SHAPE,
 }
 ThemeProvider.contextTypes = {
-  [CHANNEL]: PropTypes.func,
+  [CHANNEL_NEXT]: CONTEXT_CHANNEL_SHAPE,
 }
 
 export default ThemeProvider